.Projects/youtube_discord_bot/old_requirements.txt
discord.py
urllib3
requests
BeautifulSoup4
selenium
pynacl
youtube_dl